Added a decorator for setting a default timeout, and a decorator for setting a "key prefix"
