+++ src/Backend/DefaultTimeoutDecorated.php
@@ -10,17 +10,13 @@ namespace Matomo\Cache\Backend;
 
 use Matomo\Cache\Backend;
 
-class DefaultTimeoutDecorated implements Backend
+class DefaultTimeoutDecorated extends BaseDecorator
 {
     /**
      * @var integer
      */
     private $defaultTTL;
 
-    /**
-     * @var Backend
-     */
-    private $decorated;
 
     /**
      * Constructor.
@@ -37,37 +33,11 @@ class DefaultTimeoutDecorated implements Backend
         }
 
         $this->defaultTTL = $options['defaultTimeout'];
-        $this->decorated = $decorated;
-    }
-
-    public function doFetch($id)
-    {
-        return $this->decorated->doFetch($id);
-    }
-
-    public function doContains($id)
-    {
-        return $this->decorated->doContains($id);
+        parent::__construct($decorated);
     }
 
     public function doSave($id, $data, $lifeTime = 0)
     {
         return $this->decorated->doSave( $id, $data, $lifeTime ?: $this->defaultTTL);
     }
-
-    public function doDelete($id)
-    {
-        return $this->decorated->doDelete($id);
-    }
-
-    public function doFlush()
-    {
-        return $this->decorated->doFlush();
-    }
-
-    public function getBackend()
-    {
-        return $this->decorated;
-    }
-
 }
